+++ src/components/super/i-block/modules/activation/index.ts
@@ -176,11 +176,7 @@ export function onActivated(component: iBlock, force: boolean = false): void {
 		unsafe.componentStatus = 'beforeReady';
 	}
 
-	const needInitLoadOrReload =
-		!unsafe.isReadyOnce &&
-		force || unsafe.reloadOnActivation;
-
-	if (needInitLoadOrReload) {
+	if (unsafe.reloadOnActivation) {
 		const group = {group: 'requestSync:get'};
 		async.forEach(($a) => $a.clearAll(group).setImmediate(load, group));
 	}
@@ -220,6 +216,11 @@ export function onDeactivated(component: iBlock): void {
 
 	async.forEach(($a) => {
 		Object.entries(Namespaces).forEach(([key, namespace]) => {
+			if (unsafe.reloadOnActivation && namespace === Namespaces.request) {
+				$a.muteRequest({group: 'i-data:initLoad'});
+				return;
+			}
+
 			if (Object.isNumber(namespace) && nonMuteAsyncNamespaces[namespace]) {
 				return;
 			}