+++ CHANGELOG.md
@@ -20,6 +20,10 @@ with certain warnings. Vue uses a `get` trap within the proxy to verify the pres
 of a property in the instance. Accessing undefined properties via the `getComponentInfo` method
 during a warn or error handler will trigger infinite recursion. `core/component/engines/vue3`
 
+#### :house: Internal
+
+* Reloading now occurs only if the `reloadOnActivation` option is set to `true` `components/super/i-block/modules/activation`
+
 ## v4.0.0-beta.149 (2024-10-31)
 
 #### :rocket: New Feature